Initial Commit of SHETRAN
Compiled Linux version 4.5.2. Contains both prepare and run scripts. 4 example catchments and a readme - Foston Beck (100m), Haltwhistle (50m) and a historical SHETRAN UK simulation for 7006. 7006 is split into the files that go into the prepare script and the input_xxx files made by the prepare script that are used by the SHETRAN simulation. This is so that both the prepare and run softwares can be tested seperately.